refactor: move run code to run folder
