+++ lab/runs/tests/test_admin_change_state_action.py
@@ -10,10 +10,10 @@ from django.core.exceptions import ValidationError
 from django.test.client import RequestFactory
 from django.urls import reverse
 
-from ...admin import run_actions
-from ...admin.run import RunAdmin
-from ...models import Run
-from .. import factories
+from ...tests import factories
+from .. import admin_actions
+from ..admin import RunAdmin
+from ..models import Run
 
 
 @pytest.fixture
@@ -26,44 +26,44 @@ def changelist_request():
     return RequestFactory().get(reverse("admin:lab_run_changelist"))
 
 
-@mock.patch.object(run_actions, "validate_mandatory_fields", mock.Mock())
-@mock.patch.object(run_actions, "validate_1_method_required", mock.Mock())
-@mock.patch.object(run_actions, "validate_not_last_state", mock.Mock())
-@mock.patch.object(run_actions, "validate_execute_needs_admin", mock.Mock())
-@mock.patch.object(run_actions, "send_message", mock.Mock())
-@mock.patch.object(run_actions, "change_status", mock.Mock())
+@mock.patch.object(admin_actions, "validate_mandatory_fields", mock.Mock())
+@mock.patch.object(admin_actions, "validate_1_method_required", mock.Mock())
+@mock.patch.object(admin_actions, "validate_not_last_state", mock.Mock())
+@mock.patch.object(admin_actions, "validate_execute_needs_admin", mock.Mock())
+@mock.patch.object(admin_actions, "send_message", mock.Mock())
+@mock.patch.object(admin_actions, "change_status", mock.Mock())
 def test_change_state_calls_validators_and_changes_status(
     modeladmin, changelist_request
 ):
     queryset = [Run()]
     changelist_request.user = None
 
-    run_actions.change_state(modeladmin, changelist_request, queryset)
+    admin_actions.change_state(modeladmin, changelist_request, queryset)
 
-    run_actions.validate_mandatory_fields.assert_called_once()
-    run_actions.validate_1_method_required.assert_called_once()
-    run_actions.validate_not_last_state.assert_called_once()
-    run_actions.validate_execute_needs_admin.assert_called_once()
-    run_actions.change_status.assert_called_once()
+    admin_actions.validate_mandatory_fields.assert_called_once()
+    admin_actions.validate_1_method_required.assert_called_once()
+    admin_actions.validate_not_last_state.assert_called_once()
+    admin_actions.validate_execute_needs_admin.assert_called_once()
+    admin_actions.change_status.assert_called_once()
 
 
-@mock.patch.object(run_actions, "validate_mandatory_fields", mock.Mock())
-@mock.patch.object(run_actions, "validate_1_method_required", mock.Mock())
-@mock.patch.object(run_actions, "validate_not_last_state", mock.Mock())
+@mock.patch.object(admin_actions, "validate_mandatory_fields", mock.Mock())
+@mock.patch.object(admin_actions, "validate_1_method_required", mock.Mock())
+@mock.patch.object(admin_actions, "validate_not_last_state", mock.Mock())
 @mock.patch.object(
-    run_actions,
+    admin_actions,
     "validate_execute_needs_admin",
     mock.Mock(side_effect=ValidationError("test")),
 )
-@mock.patch.object(run_actions, "send_message", mock.Mock())
-@mock.patch.object(run_actions, "change_status", mock.Mock())
+@mock.patch.object(admin_actions, "send_message", mock.Mock())
+@mock.patch.object(admin_actions, "change_status", mock.Mock())
 def test_change_state_class_message(modeladmin, changelist_request):
     queryset = [Run()]
     changelist_request.user = None
 
-    run_actions.change_state(modeladmin, changelist_request, queryset)
+    admin_actions.change_state(modeladmin, changelist_request, queryset)
 
-    run_actions.send_message.assert_called_with(changelist_request, "test", "error")
+    admin_actions.send_message.assert_called_with(changelist_request, "test", "error")
 
 
 @pytest.mark.parametrize(
@@ -73,14 +73,14 @@ def test_change_state_class_message(modeladmin, changelist_request):
 def test_not_last_state_can_change(status):
     run = factories.RunReadyToAskExecFactory.build(status=status)
 
-    run_actions.validate_not_last_state(run)
+    admin_actions.validate_not_last_state(run)
 
 
 def test_last_state_can_change():
     run = factories.RunReadyToAskExecFactory.build(status=Run.Status.FINISHED)
 
     with pytest.raises(ValidationError):
-        run_actions.validate_not_last_state(run)
+        admin_actions.validate_not_last_state(run)
 
 
 @pytest.mark.parametrize(
@@ -90,10 +90,10 @@ def test_last_state_can_change():
 def test_status_can_move_after_new_if_fields_are_defined(status):
     run = factories.RunReadyToAskExecFactory.build(status=status)
 
-    run_actions.validate_mandatory_fields(run)
+    admin_actions.validate_mandatory_fields(run)
 
 
-@pytest.mark.parametrize("mandatory_field", run_actions.MANDATORY_FIELDS)
+@pytest.mark.parametrize("mandatory_field", admin_actions.MANDATORY_FIELDS)
 @pytest.mark.parametrize(
     "status",
     [s for s in Run.Status if s != Run.Status.FINISHED],
@@ -104,7 +104,7 @@ def test_not_new_requires_mandatory_field_to_be_defined(status, mandatory_field)
     )
 
     with pytest.raises(ValidationError):
-        run_actions.validate_mandatory_fields(run)
+        admin_actions.validate_mandatory_fields(run)
 
 
 @pytest.mark.parametrize("method_field", Run.get_method_fields())
@@ -113,21 +113,21 @@ def test_one_method_selected_allows_for_ask_exec(method_field):
         **{"status": Run.Status.CREATED, method_field.name: True}
     )
 
-    run_actions.validate_1_method_required(run)
+    admin_actions.validate_1_method_required(run)
 
 
 def test_no_method_selected_does_not_allow_for_ask_exec():
     run = factories.RunForceNoMethodFactory.build(**{"status": Run.Status.CREATED})
 
     with pytest.raises(ValidationError):
-        run_actions.validate_1_method_required(run)
+        admin_actions.validate_1_method_required(run)
 
 
 def test_member_can_ask_for_execution():
     run = factories.RunFactory.build(status=Run.Status.CREATED)
     user = factories.StaffUserFactory.build()
 
-    run_actions.validate_execute_needs_admin(user, run)
+    admin_actions.validate_execute_needs_admin(user, run)
 
 
 @pytest.mark.parametrize(
@@ -138,7 +138,7 @@ def test_admin_can_change_state(status):
     run = factories.RunFactory.build(status=status)
     user = factories.LabAdminUserFactory.build()
 
-    run_actions.validate_execute_needs_admin(user, run)
+    admin_actions.validate_execute_needs_admin(user, run)
 
 
 @pytest.mark.parametrize(
@@ -150,4 +150,4 @@ def test_member_cant_do_more_than_ask_for_execution(status):
     user = factories.StaffUserFactory.build()
 
     with pytest.raises(ValidationError):
-        run_actions.validate_execute_needs_admin(user, run)
+        admin_actions.validate_execute_needs_admin(user, run)
