+++ lab/runs/tests/test_models.py
@@ -4,7 +4,7 @@ from ...models import Run
 
 
 def test_next_action():
-    assert Run().next_status() == Run.Status.CREATED
+    assert Run().status == Run.Status.CREATED
     assert Run(status=Run.Status.CREATED).next_status() == Run.Status.ASK_FOR_EXECUTION
     assert Run(status=Run.Status.ASK_FOR_EXECUTION).next_status() == Run.Status.ONGOING
     assert Run(status=Run.Status.ONGOING).next_status() == Run.Status.FINISHED
